"Replace UserAuthForm with RegisterForm in Authentication pages and remove deprecated component"
This commit updates the authentication pages for Login, Forgot Password, and Register to use the new `RegisterForm` component instead of the deprecated `UserAuthForm` component. The `UserAuthForm` component has been removed from the project. The changes include updating the import statements and replacing the component usage in each page. The `RegisterForm` component is a modified version of the `UserAuthForm` component that is specific to the registration process.