"Add PersonalInfo and ActivityLog components to Profile page"
This commit adds the `PersonalInfo` and `ActivityLog` components to the `Profile` page. The background color of the container has also been changed to include a border with the same color. The previous code contained a hardcoded image and a "New Visit" button, which have been removed. The "New Appointment" button has been retained.